+++ FTClient.hh
@@ -7,7 +7,7 @@
 
 struct ServerConn {
     string ip;
-    int port;
+    uint16_t port;
     FTTimestampState ft_timestamp;
 };
 
@@ -15,11 +15,11 @@ class FTClient {
 private:
     std::vector<pthread_t> _threads;
     std::vector<string> _ips;
-    std::vector<int> _ports;
+    std::vector<uint16_t> _ports;
 
     static void* _send(void* param);
 
 public:
-    FTClient(std::vector<string>& ips, std::vector<int>& ports);
+    FTClient(std::vector<string>& ips, std::vector<uint16_t>& ports);
     bool send(FTTimestampState& ft_timestamp);
 };
